% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covidGenAgeCaseInf.R
\docType{data}
\name{covidGenAgeCaseInf}
\alias{covidGenAgeCaseInf}
\title{covidGenAgeCaseInf: COVID19 Infected Status(Sex/Age) of Korea
Ministry of Health and Welfare, COVID19 Infected Status(Sex/Age) of Korea}
\format{
A data frame with 2575 rows and 10 variables:
\describe{
\item{\code{X}}{integer COLUMN_DESCRIPTION}
\item{\code{confCase}}{integer COLUMN_DESCRIPTION}
\item{\code{confCaseRate}}{double COLUMN_DESCRIPTION}
\item{\code{createDt}}{character COLUMN_DESCRIPTION}
\item{\code{criticalRate}}{double COLUMN_DESCRIPTION}
\item{\code{death}}{integer COLUMN_DESCRIPTION}
\item{\code{deathRate}}{double COLUMN_DESCRIPTION}
\item{\code{gubun}}{character COLUMN_DESCRIPTION}
\item{\code{seq}}{integer COLUMN_DESCRIPTION}
\item{\code{updateDt}}{character COLUMN_DESCRIPTION}
}
}
\usage{
covidGenAgeCaseInf
}
\description{
Corona19 City/Provocation Status data obtained from Korea Ministry of Health and Welfare
}
\details{
DETAILS
data from Korea Ministry of Health and Welfare
Corona19 City/Provocation Status
http://openapi.data.go.kr/openapi/service/rest/Covid19/getCovid19GenAgeCaseInfJson
You can find detailed description for this data from
https://www.data.go.kr/data/15043377/openapi.do
}
\keyword{datasets}
