% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covidInf.R
\docType{data}
\name{covidInf}
\alias{covidInf}
\title{covidInf}
\format{
A data frame with 270 rows and 15 variables:
\describe{
\item{\code{X}}{integer COLUMN_DESCRIPTION}
\item{\code{accDefRate}}{double COLUMN_DESCRIPTION}
\item{\code{accExamCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{accExamCompCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{careCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{clearCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{createDt}}{character COLUMN_DESCRIPTION}
\item{\code{deathCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{decideCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{examCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{resutlNegCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{seq}}{integer COLUMN_DESCRIPTION}
\item{\code{stateDt}}{integer COLUMN_DESCRIPTION}
\item{\code{stateTime}}{character COLUMN_DESCRIPTION}
\item{\code{updateDt}}{character COLUMN_DESCRIPTION}
}
}
\usage{
covidInf
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
