% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidSupp.R
\docType{data}
\name{covidSupp}
\alias{covidSupp}
\title{DATASET_TITLE}
\format{
A data frame with 4273 rows and 7 variables:
\describe{
\item{\code{no}}{integer COLUMN_DESCRIPTION}
\item{\code{region}}{integer COLUMN_DESCRIPTION}
\item{\code{startup.year}}{integer COLUMN_DESCRIPTION}
\item{\code{year}}{integer COLUMN_DESCRIPTION}
\item{\code{month}}{integer COLUMN_DESCRIPTION}
\item{\code{loan.krw1mil}}{integer COLUMN_DESCRIPTION}
\item{\code{industry}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
covidSupp
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
