% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltCovidInf.R
\name{pltCovidInf}
\alias{pltCovidInf}
\title{pltCovidInf}
\usage{
pltCovidInf(CovidInf, plotopt = 1)
}
\arguments{
\item{CovidInf}{name of dataset}

\item{plotopt}{options for plot}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
plot function for CovidInf data
}
\details{
DETAILS
}
\examples{
 library(overcomeCovidKor)
 pltCovidInf(covidSupp)
}
\seealso{
\code{\link[dplyr]{group_by}}
}
