% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltcovidSupp.R
\name{pltcovidSupp}
\alias{pltcovidSupp}
\title{pltcovidSupp}
\usage{
pltcovidSupp(covidSupp, plotopt)
}
\arguments{
\item{plotopt}{options for plot}

\item{CovidInf}{name of dataset}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
plot function for CovidInf data
}
\details{
DETAILS
}
\examples{
 library(overcomeCovidKor)
 pltcovidSupp(covidSupp, 1)
}
\seealso{
\code{\link[dplyr]{group_by}}
}
