% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covidGenAgeCaseInf.R
\docType{data}
\name{covidGenAgeCaseInf}
\alias{covidGenAgeCaseInf}
\title{DATASET_TITLE}
\format{
A data frame with 2575 rows and 10 variables:
\describe{
\item{\code{X}}{integer COLUMN_DESCRIPTION}
\item{\code{confCase}}{integer COLUMN_DESCRIPTION}
\item{\code{confCaseRate}}{double COLUMN_DESCRIPTION}
\item{\code{createDt}}{character COLUMN_DESCRIPTION}
\item{\code{criticalRate}}{double COLUMN_DESCRIPTION}
\item{\code{death}}{integer COLUMN_DESCRIPTION}
\item{\code{deathRate}}{double COLUMN_DESCRIPTION}
\item{\code{gubun}}{character COLUMN_DESCRIPTION}
\item{\code{seq}}{integer COLUMN_DESCRIPTION}
\item{\code{updateDt}}{character COLUMN_DESCRIPTION}
}
}
\usage{
covidGenAgeCaseInf
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
