% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covidSidoInf.R
\docType{data}
\name{covidSidoInf}
\alias{covidSidoInf}
\title{covidSidoInf}
\format{
A data frame with 5143 rows and 16 variables:
\describe{
\item{\code{X}}{integer COLUMN_DESCRIPTION}
\item{\code{createDt}}{character COLUMN_DESCRIPTION}
\item{\code{deathCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{defCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{gubun}}{character COLUMN_DESCRIPTION}
\item{\code{gubunCn}}{character COLUMN_DESCRIPTION}
\item{\code{gubunEn}}{character COLUMN_DESCRIPTION}
\item{\code{incDec}}{integer COLUMN_DESCRIPTION}
\item{\code{isolClearCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{isolIngCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{localOccCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{overFlowCnt}}{integer COLUMN_DESCRIPTION}
\item{\code{qurRate}}{character COLUMN_DESCRIPTION}
\item{\code{seq}}{integer COLUMN_DESCRIPTION}
\item{\code{stdDay}}{character COLUMN_DESCRIPTION}
\item{\code{updateDt}}{character COLUMN_DESCRIPTION}
}
}
\usage{
covidSidoInf
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
