% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCovid.R
\name{pltCovidInf}
\alias{pltCovidInf}
\title{FUNCTION_TITLE}
\usage{
pltCovidInf(covidSupp)
}
\arguments{
\item{covidSupp}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
 library(overcomeCovidKor)
 pltCovidInf(covidSupp)
}
\seealso{
\code{\link[dplyr]{group_by}}
}
